+++ CHANGES.md
@@ -76,6 +76,8 @@ Features:
 Changes:
 - `wva`, `wvactapp` now use ClusterSite #1748
 - `conn` and `connappct` now use ClusterSite #1733
+- Migrate from requests to httpx AsyncClient client. #739
+- Migrate sync calls to async. #739
 
 Fixes:
 - fix `tenn` miss adding opinion types #1749